Move the code that creates Consul containers out of teststorage.  This allows importers of teststorage that don't need consul to run as a non-docker test.
